quote-agent-python/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
twilio==8.9.1
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.4.2
jinja2==3.1.2
httpx==0.25.0
python-multipart==0.0.6
aiosmtplib==2.0.2
aiohttp==3.8.5
# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0
psycopg2-binary==2.9.7
